Rename modules using verbs
