+++ acrosome_counter/build_model.py
@@ -6,7 +6,7 @@ from os.path import join
 from detectron2 import model_zoo
 from detectron2.config import get_cfg
 
-from .inputs import QTY_CLASSES
+from .load_dataset import QTY_CLASSES
 
 LOG_DIR = join(".", "logs")
 PRETRAINED_CHECKPOINT = model_zoo.get_checkpoint_url(
