+++ acrosome_counter/predict.py
@@ -10,7 +10,7 @@ from detectron2.engine import DefaultPredictor
 from detectron2.data import MetadataCatalog
 import pandas as pd
 
-from .inputs import MAP_NAMES
+from .load_dataset import MAP_NAMES
 from .visualize import visualize
 
 mpl.use('TkAgg')
